Service hooks and unhooks DnsCache on Win8 Service requires admin privs Clean service exit, removed Kill() Added About window, but not used yet Kitty Pix
git-svn-id: https://svn.alienseed.com:8443/svn/alienseed/trunk/dotBitNS@2711 4bf54c5a-0cb7-ff49-8398-c775ed83c514

+++ dotBitNS/Core/Program.cs
@@ -301,7 +301,8 @@ namespace dotBitNS
             if (restart)
                 Process.Start(ExePath, Arguments);
 
-            m_Process.Kill();
+            if(!IsService)
+                m_Process.Kill();
         }
 
         private static void HandleClosed()
@@ -317,6 +318,8 @@ namespace dotBitNS
                 EventSink.InvokeShutdown(new ShutdownEventArgs());
 
             Timer.TimerThread.Set();
+
+            UI.Monitor.DisableCacheEntries();
 
             Console.WriteLine("done");
         }
